requirements.txt
pyactiveresource==2.2.2
PyJWT==2.3.0
python-dotenv==0.19.2
PyYAML==6.0
ShopifyAPI==10.0.0
six==1.16.0
